Sépare Legend en composantes réutilisables
En particulier utile pour un redesign mobile en tabs, où la légende ne
serait pas réellement utilisée (seulement utile pour desktop), et le
layout utiliserait les composantes individuelles.
